requirements.txt
numpy
streamlit
pandas
numpy
plotly.express